@@ -134,9 +134,9 @@ namespace UnicornHack.Utils.MessagingECS
                     out var sortKey))
                 {
                     sortKey = referencingEntities.Where(p => p.Value == entity).Select(p => p.Key).FirstOrDefault();
-                }
+                }                
 
-                if (sortKey != default
+                if (!Equals(sortKey, default)
                     && referencingEntities.Remove(sortKey))
                 {
                     if (referencingEntities.Count == 0)
