@@ -7,7 +7,7 @@
     <div class="intro">
         <div style="margin-right: auto; margin-left: auto">
             <h1 class="text-center">UnicornHack</h1>
-            <br/>
+            <br />
             <pre class="text-left">
     \
      \
@@ -44,6 +44,11 @@
                     <button type="submit" class="form-inline btn-secondary" style="margin: auto">Play</button>
                 </div>
             </form>
+        </div>
+        <div style="display: flex; flex-wrap: wrap; width: 100%">
+            <p class="text-center" style="margin-top: 20px; width: 100%">
+                Notice: this game is currently in prototype stage and isn't expected to work correcly or be fun.
+            </p>
         </div>
     </div>
 </div>
