@@ -1,21 +1,41 @@
-﻿<!DOCTYPE html>
-<html>
+﻿@using System.IO
+@inject Microsoft.AspNetCore.Hosting.IWebHostEnvironment WebHostEnvironment
+
+<!DOCTYPE html>
+<html lang="en">
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no" />
     <title>@ViewData["Title"] - UnicornHack</title>
     <base href="~/" />
+    @*
+        manifest.json provides metadata used when your web app is added to the
+        homescreen on Android. See https://developers.google.com/web/fundamentals/engage-and-retain/web-app-manifest/
+    *@
+    <link rel="manifest" href="~/manifest.json">
+    <link rel="shortcut icon" href="~/favicon.ico">
 
+    @*In development the css is embedded*@
     <environment exclude="Development">
-        <link rel="stylesheet" href="~/dist/site.css" asp-append-version="true" />
+        <link rel="stylesheet" href="~/css/main.chunk.css" asp-append-version="true" />
     </environment>
     @Html.Raw(JavaScriptSnippet.FullScript)
 </head>
-<body style="background-color: #002b36"> @*Avoid white flash before the style loads*@
+@*Avoid white flash before the style loads*@
+<body style="background-color: #002b36">
     @RenderBody()
-
-    <script src="~/dist/vendor.js" asp-append-version="true"></script>
-    <script src="~/dist/main.js" asp-append-version="true"></script>
+    <environment exclude="Development">
+        <script src="~/js/vendors-main.chunk.js" asp-append-version="true"></script>
+        <script src="~/js/main.chunk.js" asp-append-version="true"></script>
+        <script>
+            @Html.Raw(File.ReadAllText(WebHostEnvironment.WebRootFileProvider.GetFileInfo("js/runtime~main.js").PhysicalPath))
+        </script>
+    </environment>
+    <environment include="Development">
+        <script src="~/js/vendors-main.chunk.js"></script>
+        <script src="~/js/main.chunk.js"></script>
+        <script src="~/js/runtime~main.js"></script>
+    </environment>
     @RenderSection("scripts", required: false)
 </body>
 </html>
